+++ nutzboot-core/src/main/java/org/nutz/boot/starter/WebServletFace.java
@@ -10,6 +10,10 @@ public interface WebServletFace {
     String getName();
 
     String getPathSpec();
+    
+    default String[] getPathSpecs() {
+        return new String[] {getPathSpec()};
+    }
 
     Servlet getServlet();